Expose createFusionAuth and fusionAuthKey from Vue SDK
