+++ packages/sdk-vue/src/index.ts
@@ -1,43 +1,6 @@
-import { App } from 'vue';
-import { FusionAuthConfig } from './types.ts';
-import * as components from './components';
-import { fusionAuthKey } from './injectionSymbols.ts';
-import { createFusionAuth } from './createFusionAuth';
-
-/**
- * Installation method for the FusionAuthVuePlugin.
- *
- * @category Plugin
- * @param {App} app - The Vue app instance.
- * @param {FusionAuthConfig} options - The configuration options for the plugin.
- * @throws {Error} Will throw an error if the required options are missing.
- */
-const FusionAuthVuePlugin = {
-  install(app: App, options: FusionAuthConfig) {
-    // Validate options
-    if (!options.clientId) {
-      throw new Error('clientId is required');
-    }
-
-    if (!options.serverUrl) {
-      throw new Error('serverUrl is required');
-    }
-
-    if (!options.redirectUri) {
-      throw new Error('redirectUri is required');
-    }
-
-    app.provide(fusionAuthKey, createFusionAuth(options));
-
-    // Register the components
-    Object.entries(components).forEach(([key, component]) => {
-      app.component(key, component);
-    });
-  },
-};
-
-export default FusionAuthVuePlugin;
-
 export * from './components';
 export * from './composables/useFusionAuth';
 export * from './types';
+export * from './createFusionAuth/index';
+export * from './injectionSymbols';
+export { default } from './FusionAuthVuePlugin';
